requirements.txt
aiohttp>=3.8.0
airtable-python-wrapper>=0.15.0
pydantic>=2.0.0
python-dotenv>=1.0.0
fastapi>=0.68.0
asyncpg[pool]>=0.29.0
httpx>=0.18.0
PyJWT>=2.1.0
boto3>=1.34.5
botocore>=1.34.5
psycopg2-binary>=2.9.0
pydantic-settings>=0.1.0
tenacity>=8.0.1
structlog>=21.1.0
slowapi>=0.1.4
fastapi_limiter>=0.1.0
pytz>=2023.3
factory_boy
sqlalchemy
pydantic[email]
python-multipart>=0.0.5
# aioredis>=2.0.0
redis>=4.0.0
fastapi-cache2>=0.2.0  # Ensure this line is present
pendulum>=2.1.2
setuptools